requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.19.2
Pillow>=8.0.1
matplotlib>=3.3.2